+++ configure/channel.go
@@ -2,34 +2,54 @@ package configure
 
 import (
 	"encoding/json"
-	"flag"
+	"fmt"
 	"io/ioutil"
 	"log"
 	"math/rand"
 	"sync"
 	"time"
-)
-var (
-	roomKeySaveFile  = flag.String("keyFile", "room_keys.json", "path to save room keys")
+
+	"github.com/go-redis/redis/v7"
 )
 
-var RoomKeys = LoadRoomKey(*roomKeySaveFile)
+var RoomKeys = LoadRoomKey(*GetKeyFile())
 
 var roomUpdated = false
 
-func init() {
-	rand.Seed(time.Now().UnixNano())
-	go func() {
-		for {
-			time.Sleep(15 * time.Second)
-			if roomUpdated {
-				RoomKeys.Save(*roomKeySaveFile)
-				roomUpdated = false
-			}
-		}
-	}()
-}
+var saveInFile = true
+var redisCli *redis.Client
 
+func Init() {
+	saveInFile = GetRedisAddr() == nil
+
+	rand.Seed(time.Now().UnixNano())
+	if saveInFile {
+		go func() {
+			for {
+				time.Sleep(15 * time.Second)
+				if roomUpdated {
+					RoomKeys.Save(*roomKeySaveFile)
+					roomUpdated = false
+				}
+			}
+		}()
+
+		return
+	}
+
+	redisCli = redis.NewClient(&redis.Options{
+		Addr:     *GetRedisAddr(),
+		Password: *GetRedisPwd(),
+		DB:       0,
+	})
+
+	_, err := redisCli.Ping().Result()
+	if err != nil {
+		panic(err)
+	}
+
+	log.Printf("Redis connected")
+}
 
 type RoomKeysType struct {
 	mapChanKey sync.Map
@@ -37,7 +57,7 @@ type RoomKeysType struct {
 }
 
 func LoadRoomKey(f string) *RoomKeysType {
-	result := &RoomKeysType {
+	result := &RoomKeysType{
 		mapChanKey: sync.Map{},
 		mapKeyChan: sync.Map{},
 	}
@@ -76,8 +96,24 @@ func (r *RoomKeysType) Save(f string) {
 }
 
 // set/reset a random key for channel
-func (r *RoomKeysType) SetKey(channel string) string {
-	var key string
+func (r *RoomKeysType) SetKey(channel string) (key string, err error) {
+	if !saveInFile {
+		for {
+			key = randStringRunes(48)
+			if _, err = redisCli.Get(key).Result(); err == redis.Nil {
+				err = redisCli.Set(channel, key, 0).Err()
+				if err != nil {
+					return
+				}
+
+				err = redisCli.Set(key, channel, 0).Err()
+				return
+			} else if err != nil {
+				return
+			}
+		}
+	}
+
 	for {
 		key = randStringRunes(48)
 		if _, found := r.mapKeyChan.Load(key); !found {
@@ -87,31 +123,48 @@ func (r *RoomKeysType) SetKey(channel string) string {
 		}
 	}
 	roomUpdated = true
-	return key
+	return
 }
 
-func (r *RoomKeysType) GetKey(channel string) string {
+func (r *RoomKeysType) GetKey(channel string) (newKey string, err error) {
+	if !saveInFile {
+		if newKey, err = redisCli.Get(channel).Result(); err == redis.Nil {
+			newKey, err = r.SetKey(channel)
+			log.Printf("[KEY] new channel [%s]: %s", channel, newKey)
+			return
+		}
+
+		return
+	}
+
 	var key interface{}
 	var found bool
 	if key, found = r.mapChanKey.Load(channel); found {
-		return key.(string)
-	} else {
-		newkey := r.SetKey(channel)
-		log.Printf("[KEY] new channel [%s]: %s", channel, newkey)
-		return newkey
+		return key.(string), nil
 	}
+	newKey, err = r.SetKey(channel)
+	log.Printf("[KEY] new channel [%s]: %s", channel, newKey)
+	return
 }
 
-func (r *RoomKeysType) GetChannel(key string) string {
-	channel, found := r.mapKeyChan.Load(key)
+func (r *RoomKeysType) GetChannel(key string) (channel string, err error) {
+	if !saveInFile {
+		return redisCli.Get(key).Result()
+	}
+
+	chann, found := r.mapKeyChan.Load(key)
 	if found {
-		return channel.(string)
+		return chann.(string), nil
 	} else {
-		return ""
+		return "", fmt.Errorf("%s does not exists", key)
 	}
 }
 
 func (r *RoomKeysType) DeleteChannel(channel string) bool {
+	if !saveInFile {
+		return redisCli.Del(channel).Err() != nil
+	}
+
 	key, ok := r.mapChanKey.Load(channel)
 	if ok {
 		r.mapChanKey.Delete(channel)
@@ -122,6 +175,10 @@ func (r *RoomKeysType) DeleteChannel(channel string) bool {
 }
 
 func (r *RoomKeysType) DeleteKey(key string) bool {
+	if !saveInFile {
+		return redisCli.Del(key).Err() != nil
+	}
+
 	channel, ok := r.mapKeyChan.Load(key)
 	if ok {
 		r.mapChanKey.Delete(channel)
@@ -130,7 +187,6 @@ func (r *RoomKeysType) DeleteKey(key string) bool {
 	}
 	return false
 }
-
 
 // helpers
 var letterRunes = []rune("0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")